refactor and add support for specifying score when `open survey in editor`
