+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib/src/main/java/com/microsoft/azure/toolkit/intellij/common/feedback/MonkeySurveyEditorProvider.java
@@ -1,0 +1,37 @@
+package com.microsoft.azure.toolkit.intellij.common.feedback;
+
+import com.intellij.openapi.fileEditor.FileEditor;
+import com.intellij.openapi.fileEditor.FileEditorPolicy;
+import com.intellij.openapi.fileEditor.FileEditorProvider;
+import com.intellij.openapi.project.DumbAware;
+import com.intellij.openapi.project.Project;
+import com.intellij.openapi.vfs.VirtualFile;
+
+import javax.annotation.Nonnull;
+
+public class MonkeySurveyEditorProvider implements FileEditorProvider, DumbAware {
+    public static final String TYPE = "Microsoft.Customer_Survey";
+
+    @Override
+    public boolean accept(@Nonnull Project project, @Nonnull VirtualFile virtualFile) {
+        return virtualFile.getFileType().getName().equals(getEditorTypeId());
+    }
+
+    @Nonnull
+    @Override
+    public FileEditor createEditor(@Nonnull Project project, @Nonnull VirtualFile virtualFile) {
+        return new MonkeySurveyEditor(project, virtualFile);
+    }
+
+    @Nonnull
+    @Override
+    public String getEditorTypeId() {
+        return TYPE;
+    }
+
+    @Nonnull
+    @Override
+    public FileEditorPolicy getPolicy() {
+        return FileEditorPolicy.HIDE_DEFAULT_EDITOR;
+    }
+}
