+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib/src/main/resources/META-INF/azure-intellij-plugin-lib.xml
@@ -14,7 +14,7 @@
                     id="Azure Streaming Log" canCloseContents="true"/>
         <webHelpProvider implementation="com.microsoft.azure.toolkit.intellij.common.help.AzureWebHelpProvider"/>
         <applicationService serviceImplementation="com.microsoft.azure.toolkit.intellij.common.settings.IntellijStore"/>
-        <fileEditorProvider implementation="com.microsoft.azure.toolkit.intellij.common.feedback.ProvideFeedbackEditorProvider"/>
+        <fileEditorProvider implementation="com.microsoft.azure.toolkit.intellij.common.feedback.MonkeySurveyEditorProvider"/>
     </extensions>
     <actions>
         <action id="Actions.WhatsNew"
